refactor: consolidate mimeTypeUtil into src/utils
Move MIME type mapping to src/utils/mimeTypeUtil.ts with both
getMimeType (ext→mime) and getExtension (mime→ext) derived from a
single map. Browser test helpers now import from the shared util.
